@@ -2,11 +2,27 @@
     "name": "alkaruno/phpframework",
     "description": "Yet another PHP Framework",
     "require": {
-        "php": ">=5.4"
+        "php": ">=7.0",
+        "ext-json": "*",
+        "ext-pdo": "*"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "~5",
+        "squizlabs/php_codesniffer": "^2",
+        "phpmd/phpmd": "~2",
+        "smarty/smarty": "^3.1",
+        "twig/twig": "~1.0",
+        "swiftmailer/swiftmailer": "^5.4"
     },
     "autoload": {
         "psr-4": {
-            "Xplosio\\PhpFramework\\": "src"
+            "framework\\": "src"
         }
+    },
+    "scripts": {
+        "test": "phpunit"
+    },
+    "config": {
+        "sort-packages": true
     }
 }
