@@ -1,75 +1,71 @@
 <?php
 
-namespace Xplosio\PhpFramework;
+namespace framework;
 
 use Exception;
+use framework\web\Request;
 
 class Error
 {
     public static function handle()
     {
-        $args = func_get_args();
+        $args = \func_get_args();
 
-        if (count($args) == 5) {
-            list($code, $text, $file, $line, $info) = $args;
-            $info = null;
-            if (is_array($info)) {
+        if (\count($args) === 5) {
+            list(, $text, $file, $line, $info) = $args;
+            if (\is_array($info)) {
                 $info = print_r($info, true);
             }
             $code = 500;
         } else {
-            /**
-             * @var Exception $e
-             */
-            $e = $args[0];
-            $code = $e->getCode();
-            $text = $e->getMessage();
-            $file = $e->getFile();
-            $line = $e->getLine();
-            $info = $e->getTraceAsString();
+            /** @var Exception $exception */
+            $exception = $args[0];
+            $code = $exception->getCode();
+            $text = $exception->getMessage();
+            $file = $exception->getFile();
+            $line = $exception->getLine();
+            $info = $exception->getTraceAsString();
         }
 
-        $errors = array(
+        $errors = [
             401 => 'Unauthorized',
             403 => 'Forbidden',
             404 => 'Not Found',
             500 => 'Internal Server Error'
-        );
+        ];
 
-        if (!isset($errors[$code])) {
+        if (!array_key_exists($code, $errors)) {
             $code = 500;
         }
 
         if (!headers_sent()) {
-            header('HTTP/1.1 ' . isset($errors[$code]) ? $errors[$code] : '', true, $code);
+            header(sprintf('HTTP/1.0 %d %s', $code, $errors[$code]), true, $code);
         }
 
         $message = "{$code} {$text} {$file}:{$line}";
-        if ($info != null) {
+        if ($info !== null) {
             $message .= "\n{$info}";
         }
 
-        $data = array(
+        Logger::error($message);
+
+        $data = [
             'code' => $code,
             'title' => $errors[$code],
             'message' => $message,
-            'debug' => isset(App::$config['env']['debug']) && App::$config['env']['debug']
-        );
+            'debug' => App::getConfigValue('debug', false)
+        ];
 
-        Logger::error($message);
+        $errorView = App::getConfigValue(['views', 'error'], null);
+        $viewsPath = App::getConfigValue(['views', 'views_path'], '../app/views');
 
-        if (isset(App::$config['errorView']) && is_readable(App::$config['views_path'] . DIRECTORY_SEPARATOR . App::$config['errorView'])) {
-            /** @var $request Request */
-            $request = App::$request;
-            foreach ($data as $key => $value) {
-                $request->set($key, $value);
-            }
-            App::showView(App::$config['errorView'], $request->getData());
+        if ($errorView !== null && is_readable($viewsPath . DIRECTORY_SEPARATOR . $errorView)) {
+            echo App::render($errorView, array_merge($data, App::$request->getAttributes()));
         } else {
-            extract($data);
-            include App::$folder . '/views/error.php';
+            extract($data, EXTR_OVERWRITE);
+            include App::getPath('@framework/views/error.php');
         }
 
-        exit;
+        exit(1);
     }
 }
