@@ -1,13 +1,27 @@
 <?php
 
-namespace Xplosio\PhpFramework;
+namespace framework;
+
+use framework\db\Db;
 
 class Validator
 {
+    const TYPE_REQUIRED = 'required';
+    const TYPE_LENGTH = 'length';
+    const TYPE_REGEXP = 'regexp';
+    const TYPE_INT = 'int';
+    const TYPE_MAIL = 'mail';
+    const TYPE_UNIQUE = 'unique';
+    const TYPE_CONFIRM = 'confirm';
+    const TYPE_IN = 'in';
+
+    const TYPE_CUSTOM = 'custom';
+
     private $data;
     private $autotrim;
-    private $fields = array();
-    private $errors = array();
+
+    private $fields = [];
+    private $errors = [];
 
     public function __construct($autotrim = true)
     {
@@ -16,72 +30,68 @@ class Validator
 
     public function add($field)
     {
-        $validatorRule = new ValidatorRule();
+        $validatorRule = new ValidatorRules();
         $this->fields[$field] = $validatorRule;
         return $validatorRule;
     }
 
-    public function validate($data = null)
+    public function validate(array $data = null)
     {
-        $data = $data != null ? $data : $_POST;
+        mb_internal_encoding('utf-8');
+
+        $this->data = $data !== null ? $data : $_POST;
+        $this->errors = [];
 
         if ($this->autotrim) {
-            $this->data = array();
-            foreach ($data as $name => $value) {
-                $this->data[$name] = is_string($value) ? trim($value) : $value;
-            }
-        } else {
-            $this->data = $data;
+            $this->data = array_map(function ($value) {
+                return is_string($value) ? trim($value) : $value;
+            }, $this->data);
         }
 
+        $validators = $this->getValidators();
+
         /**
-         * @var ValidatorRule $validatorRule
+         * @var ValidatorRules $rules
          */
-        foreach ($this->fields as $fieldName => $validatorRule) {
+        foreach ($this->fields as $fieldName => $rules) {
 
-            $value = isset($this->data[$fieldName]) ? $this->data[$fieldName] : '';
+            $value = array_key_exists($fieldName, $this->data) ? $this->data[$fieldName] : '';
 
-            foreach ($validatorRule->getRules() as $rule) {
-
-                if (isset($this->errors[$fieldName])) {
-                    continue;
-                }
+            foreach ($rules->getRules() as $rule) {
 
                 list($type, $error, $params) = $rule;
 
-                if ($value == '' && $type != 'required' && !$validatorRule->hasCustomValidators()) {
+                if ($value === '' && $type !== self::TYPE_REQUIRED && !$rules->isHasCustomValidator()) {
                     continue;
                 }
 
-                if ($type == 'custom') {
-                    $error = call_user_func_array($params[0], array($this->data[$fieldName]));
-                    if ($error != null) {
+                if ($type === self::TYPE_CUSTOM) {
+                    $error = call_user_func_array($params[0], [$value]);
+                    if ($error !== null) {
                         $this->errors[$fieldName] = $error;
+                        break;
                     }
                 } else {
-                    if (call_user_func_array(array($this, 'validate' . ucfirst($type)), array($value, $params, $this->data)) !== true) {
+
+                    $validator = array_key_exists($type, $validators) ? $validators[$type] : null;
+                    Assert::notNull($validator, 'Not found validator for type: ' . $type);
+
+                    if (call_user_func_array($validator, [$value, $params, $this->data]) !== true) {
                         $this->errors[$fieldName] = $error;
+                        break;
                     }
                 }
             }
         }
 
-        return count($this->errors) == 0;
+        return count($this->errors) === 0;
     }
 
     public function getData($emptyToNull = true)
     {
-        $data = $this->data;
-
-        if ($emptyToNull) {
-            foreach ($data as &$row) {
-                if ($row == '') {
-                    $row = null;
-                }
-            }
-        }
-
-        return $data;
+        return array_map(function ($value) use ($emptyToNull) {
+            return $value === '' && $emptyToNull ? null : $value;
+        }, $this->data);
     }
 
     public function getErrors()
@@ -89,89 +99,56 @@ class Validator
         return $this->errors;
     }
 
-    /* validators: возвращают true, если ошибки нет */
-
-    private function validateRequired($value, $params)
+    private function getValidators()
     {
-        return $value != '';
-    }
-
-    private function validateLength($value, $params)
-    {
-        mb_internal_encoding('utf-8');
-        $length = mb_strlen($value);
-
-        if (count($params) != 2) {
-            return false;
-        }
-
-        if ($length < $params[0] || $length > $params[1]) {
-            return false;
-        }
-
-        return true;
-    }
-
-    private function validateRegexp($value, $params)
-    {
-        return preg_match('/' . $params[0] . '/u', $value) ? true : false;
-    }
-
-    private function validateMail($value, $params)
-    {
-        if ($params[0]) {
-            $pattern = '/^([^@\\s]+)@((?:[-a-z0-9]+\\.)+[a-z]{2,})$/i';
-        } else {
-            $pattern = '/^([^\\x00-\\x20\\x22\\x28\\x29\\x2c\\x2e\\x3a-\\x3c\\x3e\\x40\\x5b-'
-                    . '\\x5d\\x7f-\\xff]+|\\x22([^\\x0d\\x22\\x5c\\x80-\\xff]|\\x5c\\x00-'
-                    . '\\x7f)*\\x22)(\\x2e([^\\x00-\\x20\\x22\\x28\\x29\\x2c\\x2e\\x3a-'
-                    . '\\x3c\\x3e\\x40\\x5b-\\x5d\\x7f-\\xff]+|\\x22([^\\x0d\\x22\\x5c\\x80'
-                    . '-\\xff]|\\x5c\\x00-\\x7f)*\\x22))*\\x40([^\\x00-\\x20\\x22\\x28\\x29'
-                    . '\\x2c\\x2e\\x3a-\\x3c\\x3e\\x40\\x5b-\\x5d\\x7f-\\xff]+|\\x5b([^'
-                    . '\\x0d\\x5b-\\x5d\\x80-\\xff]|\\x5c\\x00-\\x7f)*\\x5d)(\\x2e([^\\x00-'
-                    . '\\x20\\x22\\x28\\x29\\x2c\\x2e\\x3a-\\x3c\\x3e\\x40\\x5b-\\x5d\\x7f-'
-                    . '\\xff]+|\\x5b([^\\x0d\\x5b-\\x5d\\x80-\\xff]|\\x5c\\x00-\\x7f)*'
-                    . '\\x5d))*$/';
-        }
-        return !preg_match($pattern, $value) ? false : true;
-    }
-
-    private function validateUnique($value, $params, $data)
-    {
-        $id = isset($data['id']) ? $data['id'] : null;
-
-        return Db::getValue('SELECT COUNT(1) FROM `' . $params[0] . '` WHERE ' . ($id !== null ? ' id != ' . $id . ' AND ' : '') . ' `' . $params[1] . '` = ?', $value) == 0;
-    }
-
-    private function validateHasInDatabase($value, $params, $data)
-    {
-        $id = isset($data['id']) ? $data['id'] : null;
-
-        return Db::getValue('SELECT COUNT(1) FROM `' . $params[0] . '` WHERE ' . ($id !== null ? ' id != ' . $id . ' AND ' : '') . ' `' . $params[1] . '` = ?', $value) != 0;
-    }
-
-    private function validateConfirm($value, $params, $data)
-    {
-        return $value == $data[$params[0]];
+        return [
+            self::TYPE_REQUIRED => function ($value) {
+                return $value !== '';
+            },
+            self::TYPE_LENGTH => function ($value, $params) {
+                $length = mb_strlen($value);
+                return $length >= $params[0] && $length <= $params[1];
+            },
+            self::TYPE_REGEXP => function ($value, $params) {
+                return preg_match('/' . $params[0] . '/u', $value) ? true : false;
+            },
+            self::TYPE_INT => function ($value) {
+                return filter_var($value, FILTER_VALIDATE_INT) !== false;
+            },
+            self::TYPE_MAIL => function ($value) {
+                return filter_var($value, FILTER_VALIDATE_EMAIL) !== false;
+            },
+            self::TYPE_UNIQUE => function ($value, $params, $data) {
+                $id = array_key_exists('id', $data) && is_numeric($data['id']) ? (int)$data['id'] : 0;
+                return Db::getValue("SELECT COUNT(*) FROM `$params[0]` WHERE id != ? AND `$params[1]` = ?", $id, $value) === 0;
+            },
+            self::TYPE_CONFIRM => function ($value, $params, $data) {
+                return $value === $data[$params[0]];
+            },
+            self::TYPE_IN => function ($value, $params) {
+                list($values, $strict) = $params;
+                return in_array($value, (array)$values, $strict);
+            }
+        ];
     }
 }
 
-class ValidatorRule
+class ValidatorRules
 {
-    private $rules = array();
+    private $rules = [];
     private $required = false;
-    private $hasCustomValidators = false;
+    private $hasCustomValidator = false;
 
     public function required($error)
     {
-        $this->rules[] = array('required', $error, array());
+        $this->rules[] = [Validator::TYPE_REQUIRED, $error, []];
         $this->required = true;
         return $this;
     }
 
     public function length($min, $max, $error)
     {
-        $this->rules[] = array('length', $error, array($min, $max));
+        $this->rules[] = [Validator::TYPE_LENGTH, $error, [$min, $max]];
         return $this;
     }
 
@@ -185,48 +162,56 @@ class ValidatorRule
         return $this->length(null, $size, $error);
     }
 
+    public function int($error)
+    {
+        $this->rules[] = [Validator::TYPE_INT, $error, []];
+        return $this;
+    }
+
     public function regexp($regex, $error)
     {
-        $this->rules[] = array('regexp', $error, array($regex));
+        $this->rules[] = [Validator::TYPE_REGEXP, $error, [$regex]];
         return $this;
     }
 
     public function mail($error, $strict = false)
     {
-        $this->rules[] = array('mail', $error, array($strict));
+        $this->rules[] = [Validator::TYPE_MAIL, $error, [$strict]];
         return $this;
     }
 
     public function unique($table, $field, $error)
     {
-        $this->rules[] = array('unique', $error, array($table, $field));
-        return $this;
-    }
-
-    public function hasInDatabase($table, $field, $error)
-    {
-        $this->rules[] = array('hasInDatabase', $error, array($table, $field));
+        $this->rules[] = [Validator::TYPE_UNIQUE, $error, [$table, $field]];
         return $this;
     }
 
     public function confirm($field, $error)
     {
-        $this->rules[] = array('confirm', $error, array($field));
+        $this->rules[] = [Validator::TYPE_CONFIRM, $error, [$field]];
+        return $this;
+    }
+
+    public function in($values, $error, $strict = false)
+    {
+        $this->rules[] = [Validator::TYPE_IN, $error, [$values, $strict]];
         return $this;
     }
 
     public function custom($handler)
     {
-        $this->rules[] = array('custom', null, array($handler));
-        $this->hasCustomValidators = true;
+        $this->rules[] = [Validator::TYPE_CUSTOM, null, [$handler]];
+        $this->hasCustomValidator = true;
         return $this;
     }
 
+    /** @deprecated */
     public function addRule($type, $error, $data)
     {
-        $this->rules[] = array($type, $error, array($data));
+        $this->rules[] = [$type, $error, [$data]];
     }
 
+    /** @deprecated */
     public function setRules($rules)
     {
         $this->rules = $rules;
@@ -237,13 +222,8 @@ class ValidatorRule
         return $this->rules;
     }
 
-    public function isRequired()
+    public function isHasCustomValidator()
     {
-        return $this->required;
-    }
-
-    public function hasCustomValidators()
-    {
-        return $this->hasCustomValidators;
+        return $this->hasCustomValidator;
     }
 }
