@@ -1,93 +1,113 @@
 <?php
 
-namespace Xplosio\PhpFramework;
-
-use ReflectionClass;
+namespace framework;
 
 class Logger
 {
-    const EMERGENCY = 'emergency';
-    const ALERT     = 'alert';
-    const CRITICAL  = 'critical';
-    const ERROR     = 'error';
-    const WARNING   = 'warning';
-    const NOTICE    = 'notice';
-    const INFO      = 'info';
-    const DEBUG     = 'debug';
+    public const TRACE = 0;
+    public const DEBUG = 1;
+    public const INFO = 2;
+    public const WARN = 3;
+    public const ERROR = 4;
 
-    public function emergency($message, array $context = null)
+    private static $levels = [
+        self::TRACE => 'TRACE',
+        self::DEBUG => 'DEBUG',
+        self::INFO => 'INFO',
+        self::WARN => 'WARN',
+        self::ERROR => 'ERROR'
+    ];
+
+    private const PARAM_MASK = '{}';
+
+    public static function log($level = self::INFO, $message, ...$params)
     {
-        self::log(self::EMERGENCY, $message, $context);
-    }
-
-    public static function alert($message, array $context = null)
-    {
-        self::log(self::ALERT, $message, $context);
-    }
-
-    public static function critical($message, array $context = null)
-    {
-        self::log(self::CRITICAL, $message, $context);
-    }
-
-    public static function error($message, array $context = null)
-    {
-        self::log(self::ERROR, $message, $context);
-    }
-
-    public static function warning($message, array $context = null)
-    {
-        self::log(self::WARNING, $message, $context);
-    }
-
-    public static function notice($message, array $context = null)
-    {
-        self::log(self::NOTICE, $message, $context);
-    }
-
-    public static function info($message, array $context = null)
-    {
-        self::log(self::INFO, $message, $context);
-    }
-
-    public static function debug($message, array $context = null)
-    {
-        self::log(self::DEBUG, $message, $context);
-    }
-
-    public static function log($level, $message, array $context = null)
-    {
-        $loggerLevel = isset(App::$config['logging']['level']) ? App::$config['logging']['level'] : self::WARNING;
-
-        $class = new ReflectionClass('\\Xplosio\\PhpFramework\\Logger');
-        foreach ($class->getConstants() as $const => $value) {
-            if ($value == $level) {
-                self::internalLog($level, $message, $context);
-                return;
-            }
-            if ($value == $loggerLevel) {
-                return;
-            }
+        if ($level < self::getLevel()) {
+            return;
         }
-    }
 
-    private static function internalLog($level, $message, $context)
-    {
         if (is_array($message)) {
             $message = print_r($message, true);
         }
 
-        $message .= is_array($context) ? ': ' . print_r($context, true) : PHP_EOL;
+        if (is_string($message) && !empty($params)) {
+            $message = self::getMessageWithParams($message, ...$params);
+        }
+
+        if (is_object($message)) {
+            ob_start();
+            var_dump($message);
+            $message = ob_get_clean();
+        }
 
         $filename = sprintf('../app/logs/%s.log', date('Y-m-d'));
         if (!is_dir(dirname($filename))) {
-            mkdir(dirname($filename), 0777, true);
+            @mkdir(dirname($filename), 0777, true);
         }
 
-        file_put_contents(
-            $filename,
-            sprintf('%s %s %s', date('Y.m.d H:i:s'), strtoupper($level), $message),
-            FILE_APPEND
-        );
+        $message = date('Y.m.d H:i:s') . ' [' . self::$levels[$level] . '] ' . $message;
+        file_put_contents($filename, $message . PHP_EOL, FILE_APPEND);
+    }
+
+    public static function trace($message, ...$params)
+    {
+        self::log(self::TRACE, $message, ...$params);
+    }
+
+    public static function debug($message, ...$params)
+    {
+        self::log(self::DEBUG, $message, ...$params);
+    }
+
+    public static function info($message, ...$params)
+    {
+        self::log(self::INFO, $message, ...$params);
+    }
+
+    public static function warn($message, ...$params)
+    {
+        self::log(self::WARN, $message, ...$params);
+    }
+
+    public static function error($message, ...$params)
+    {
+        self::log(self::ERROR, $message, ...$params);
+    }
+
+    public static function getMessageWithParams($message, ...$params)
+    {
+        foreach ($params as $param) {
+            if (($pos = strpos($message, self::PARAM_MASK)) !== false) {
+                if (is_array($param) || is_object($param)) {
+                    $param = json_encode($param);
+                }
+                $message = implode('', [
+                    substr($message, 0, $pos),
+                    $param,
+                    substr($message, $pos + strlen(self::PARAM_MASK))
+                ]);
+            }
+        }
+
+        return $message;
+    }
+
+    private static $level;
+
+    private static function getLevel()
+    {
+        if (self::$level === null) {
+            self::$level = self::WARN;
+            $levelName = App::getConfigValue(['logging', 'level']);
+            if ($levelName !== null) {
+                foreach (self::$levels as $level => $name) {
+                    if (strcasecmp($name, $levelName) === 0) {
+                        self::$level = $level;
+                        break;
+                    }
+                }
+            }
+        }
+        return self::$level;
     }
 }
